Benchmark t-digest's steady state.
First values are very cheap to add as the summary is almost empty. The steady
state is more interesting to benchmark since the summary contains more values
so it takes more time to find the right place where to insert the new value,
and t-digest needs to take more decisions in order to check whether it needs
to insert a new centroid or merge the new value with an existing centroid.